refactor(mcp): extract transport-agnostic engine.ts from server.ts
Split server.ts into:
- engine.ts — owns SessionState + all handleX functions + handleTool, the
  JSON-in/JSON-out entry point into the IR. No transport.
- server.ts — MCP presentation (TOOLS/RESOURCES/prompts/docs) + the MCP SDK
  transport wiring; imports { session, handleTool } from engine.

No behavior change: handleTool is byte-identical, server.ts still spawns as the
MCP server. This decouples the IR logic from the MCP transport so a second
transport (ir_service.ts, plain JSON-RPC over stdio) can reuse handleTool
without launching an MCP server — the first step of the Lean-front-door /
Turnstile integration.

Verified: tsc --noEmit clean (0 errors); bun test mcp/ — 67 pass / 0 fail.
